requirements.txt
numpy
torch
pillow
datasets
transformers
